# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sports World EPS",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api_eps/public/"
        }
    ],
    "paths": {
...
        "/api/crm2/v1/auth": {
            "post": {
                "tags": [
                    "USUARIO"
                ],
                "summary": "USUARIO",
                "operationId": "6b4fb45af8cc630a9f62f98353da25f1",
                "requestBody": {
                    "$ref": "#/components/requestBodies/LoginRequest"
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/crm2/v1/getTrainers/{idClub}": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "operationId": "51110cc27f7307613e98b31c4c092570",
                "parameters": [
                    {
                        "name": "idClub",
                        "in": "path",
                        "schema": {
                            "type": "integer",
                            "default": 75
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/crm2/v1/search": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "operationId": "d18a15ed224a36531e118372e0443748",
                "parameters": [
                    {
                        "name": "tipo",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "SOCIO"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "luis"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/crm2/v1/changeClubBase/{idClub}": {
            "put": {
                "tags": [
                    "Trainers"
                ],
                "operationId": "de839766407cd0ddf6ff24cac6a2136d",
                "parameters": [
                    {
                        "name": "idClub",
                        "in": "path",
                        "schema": {
                            "type": "integer",
                            "default": 75
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "requestBodies": {
...
            "LoginRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "username": {
                                    "type": "string",
                                    "example": ""
                                },
                                "password": {
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "secret-key",
                "in": "header"
            }
        }
    }
}